Fisher-Yates Shuffle Algorithm implementation for server and client selection
Used the boost libraries to implement the Fisher-Yates shuffle to
randomly select the nodes which will be servers and clients.

Improved the scenario by adding the command line arguments discussed by
the group, including --networks --clients --servers and --contentsize.

Also experimented with tuples and vectors to make the assignation of
nodes as easy as possible.