Alpha version of disaster-tcp.cc
Using the information we obtained in disaster-tcp-test.cc, I have
created the first version to test disaster-tcp.cc. As discussed we use
the nms-campus topology have a set number of servers and clients and
obtain the information from the scenario.

I have tested varying the parameters and things seem to be working
alright.

If all goes well, this will become the code to continue in the master
branch.

From here on in, I will concentrate on obtaining graphs for the TCP/IP
simulation.

Cheers.