R graphs, run-sim.sh and disaster-tcp.cc final experimental versions
These are the final modifications made to the TCP/IP scenario to start
running the experiments. The R scripts will probably need to be changed to
take into account the new amount of information that run-sim.sh will generate.

In any case with this build, I will run the following scenarios:
1-3 networks, 1-250 clients per network, always 1 server, random geometric
10MB average Youtube style content size, 10 times.

This will be fun.

Cheers.