R script output directory function
It came to my attention that one might need the output directory to be
somewhere else other than the current directory. I have added the -o option to
specify a directory to output the graphs. It has no checking, so one must be
careful to specify a location that actually exists.