AP Handover SmartFlooding NDN Lan
This culminates the scenario building for my paper. Using SmartFlooding the
scenario works well and should be the way to improve. All in all the
SmartFlooding is probably the best and I don't know any way to make it better
at the moment.

We shall see what I come up with tomorrow. Cheers.